Made node_types, inter_cfg_helper and base_cfg_helper files, Removed intraprocedural, changed {} to dict.
